requirements.txt
streamlit
fastapi[all]
pydantic
numpy
pandas
uvicorn
pillow
biopython
requests
fpdf
matplotlib
plotly
scikit-learn==1.6.1
catboost
lightgbm
xgboost
huggingface_hub
# itertools
